\name{irisdat}
\alias{irisdat}
\docType{data}
\title{
irisdat
}
\description{
matrix of 150 elements on 4 features, extracted from iris standard data.frame (4 first variables). See reference.
}
\format{
  The format is:
 num [1:150, 1:4] 5.1 4.9 4.7 4.6 5 5.4 4.6 5 4.4 4.9 ...
}
\references{
Fisher, R. A. (1936) _The use of multiple measurements in taxonomic problems_, Annals of Eugenics, Volume 7, Part II, Pages 179-188.
}
\examples{
plot(irisdat)
}
\keyword{datasets}

