\name{isodat}
\alias{isodat}
\docType{data}
\title{
isolet
}
\description{
matrix of 1500 elements on 617 features. Vowels taken from the isolet UCI data set.
}
\format{
  The format is:
 num [1:1500, 1:617] -0.4394 -0.4348 -0.3928 -0.5494 -0.4550 -0.4804 -0.3656 -0.5154 -0.1836 -0.2844 ...
}
\examples{
plot(isodat)
}
\keyword{datasets}

