\name{synth1}
\alias{synth1}
\docType{data}
\title{
synth1
}
\description{
matrix of 298 elements on 2 features. Synthetic data set, see reference.
}
\format{
  The format is:
 num [1:298, 1:2] 0.4550112 0.4475910 0.4357260 0.4640844 0.4460929 0.4528965 0.4542374 0.4690467 0.4396205 0.4610803 ...
}
\references{
Zelnik-Manor, L. and Perona, P. (2004). Self-tuning spectral clustering. Neural Information Processing Symposium.
}
\examples{
plot(synth1)
}
\keyword{datasets}

