\name{synth2}
\alias{synth2}
\docType{data}
\title{
synth2
}
\description{
matrix of 302 elements on 2 features. Synthetic data set, see reference.
}
\format{
  The format is:
 num [1:302, 1:2] 0.6988742  0.9332699 -0.1523060  0.5511859  0.8133928  0.5597358  0.7980163  0.1351562  0.2319774  0.9182794 ...
}
\references{
Zelnik-Manor, L. and Perona, P. (2004). Self-tuning spectral clustering. Neural Information Processing Symposium.
}
\examples{
plot(synth2)
}
\keyword{datasets}

