\name{synth3}
\alias{synth3}
\docType{data}
\title{
synth3
}
\description{
matrix of 265 elements on 2 features. Synthetic data set, see reference.
}
\format{
  The format is:
 num [1:265, 1:2] 0.1666711 0.1630728 0.1680026 0.1471314 0.1814486 0.1667588 0.1567425 0.1613646 0.1677744 0.1515392 ...
}
\references{
Zelnik-Manor, L. and Perona, P. (2004). Self-tuning spectral clustering. Neural Information Processing Symposium.
}
\examples{
plot(synth3)
}
\keyword{datasets}

