\name{synth4}
\alias{synth4}
\docType{data}
\title{
synth4
}
\description{
matrix of 621 elements on 2 features. Synthetic data set, see reference.
}
\format{
  The format is:
 num [1:621, 1:2] 1.084640271 0.414354482 0.628829245 0.246931608 0.160063275 0.078559001 0.002920947 0.103797589 1.039537189 0.045987087 ...
}
\references{
Zelnik-Manor, L. and Perona, P. (2004). Self-tuning spectral clustering. Neural Information Processing Symposium.
}
\examples{
plot(synth4)
}
\keyword{datasets}

