\name{synth5}
\alias{synth5}
\docType{data}
\title{
synth5
}
\description{
matrix of 511 elements on 2 features. Synthetic data set, see reference.
}
\format{
  The format is:
 num [1:511, 1:2] 0.09884874 0.10504572 0.11124270 0.11743968 0.12363666 0.12983365 0.13603063 0.14222761 0.14842459 0.15462157 ...
}
\references{
Zelnik-Manor, L. and Perona, P. (2004). Self-tuning spectral clustering. Neural Information Processing Symposium.
}
\examples{
plot(synth5)
}
\keyword{datasets}

