\name{synth6}
\alias{synth6}
\docType{data}
\title{
synth6
}
\description{
matrix of 237 elements on 2 features. Synthetic data set, see reference.
}
\format{
  The format is:
 num [1:237, 1:2] 0.7479543 0.7539890 0.7473104 0.7202754 0.7134525 0.6835628 0.6653153 0.6258207 0.6194835 0.5892482 ...
}
\references{
Zelnik-Manor, L. and Perona, P. (2004). Self-tuning spectral clustering. Neural Information Processing Symposium.
}
\examples{
plot(synth6)
}
\keyword{datasets}

